split forwarding rs
