+++ examples/grafana/provisioning/dashboards/huginn.yml
@@ -5,6 +5,6 @@ providers:
     folder: Huginn
     type: file
     disableDeletion: true
-    updateIntervalSeconds: 15
+    updateIntervalSeconds: 5
     options:
       path: /var/lib/grafana/dashboards
