+++ huginn-proxy-lib/src/proxy/mod.rs
@@ -5,12 +5,13 @@ pub mod forwarding;
 pub mod handler;
 pub mod http_result;
 pub mod reload;
+pub mod router;
 pub mod server;
 pub mod synthetic_response;
 pub mod transport;
-
 pub use client_pool::ClientPool;
 pub use context::{RequestContext, SecurityContext};
-pub use forwarding::{determine_http_version, find_backend_config, pick_route};
+pub use forwarding::{determine_http_version, find_backend_config};
 pub use http_result::HttpError;
+pub use router::pick_route;
 pub use server::run;
